Really clean it up
