+++ .prettierignore
@@ -4,20 +4,19 @@ __pycache__/
 .cargo/
 .github/
 .vscode/
-compiler/
-jupyterlab/.pytest_cache/
-jupyterlab/lib/
-jupyterlab/.venv/
-jupyterlab/qsharp_jupyterlab/labextension/
-library/
-npm/dist/
-npm/lib/
-npm/src/*.generated.ts
-pip/.venv/
-pip/.pytest_cache/
-pip/src/**/*.html
-playground/public/libs/
-samples/
-target/
-vscode/out/
-wasm/
+.venv/
+/compiler/
+/jupyterlab/.pytest_cache/
+/jupyterlab/lib/
+/jupyterlab/qsharp_jupyterlab/labextension/
+/library/
+/npm/dist/
+/npm/lib/
+/npm/src/*.generated.ts
+/pip/.pytest_cache/
+/pip/src/**/*.html
+/playground/public/libs/
+/samples/
+/target/
+/vscode/out/
+/wasm/
