+++ npm/src/compiler/worker-browser.ts
@@ -3,17 +3,10 @@
 
 import * as wasm from "../../lib/web/qsc_wasm.js";
 import { log } from "../log.js";
-import { invokeWorkerMethod } from "../worker-common.js";
 import { Compiler } from "./compiler.js";
-import {
-  ICompilerMethodsOnly,
-  getWorkerEventHandlers
-} from "./worker-common.js";
+import { createCompilerDispatcher } from "./worker-proxy.js";
 
-// Used to sent messages back to the client when events occur during request processing
-const evtTarget = getWorkerEventHandlers(self.postMessage);
-
-let compiler: Compiler | null = null;
+let invokeCompiler: ReturnType<typeof createCompilerDispatcher> | null = null;
 
 // This export should be assigned to 'self.onmessage' in a WebWorker
 export function messageHandler(e: MessageEvent) {
@@ -26,18 +19,24 @@ export function messageHandler(e: MessageEvent) {
 
   switch (data.type) {
     case "init":
-      log.setLogLevel(data.qscLogLevel);
-      wasm.initSync(data.wasmModule);
-      compiler = new Compiler(wasm);
+      {
+        log.setLogLevel(data.qscLogLevel);
+        wasm.initSync(data.wasmModule);
+        const compiler = new Compiler(wasm);
+        invokeCompiler = createCompilerDispatcher(
+          self.postMessage.bind(self),
+          compiler
+        );
+      }
       break;
     default:
-      if (!compiler) {
+      if (!invokeCompiler) {
         log.error(
           `Received message before the compiler was initialized: %o`,
           data
         );
       } else {
-        invokeWorkerMethod(data, compiler as ICompilerMethodsOnly, self.postMessage, evtTarget);
+        invokeCompiler(data);
       }
   }
 }
